destination picker ui fix, refactored the location picker by using existing component
